requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
itsdangerous==2.2.0
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.4
